Make "many" & "lots of" non-monotone in 1st arg
This matches the interpretation of FraCaS people.

Side note: now all generalized quantifiers are implemented as Relations.
We achieve a 90% accuracy on the first section of FraCaS corpus. The
errors are: 030, 031, 032, 046, 048, 049, 058, 069. Among them, 049 and
058 are due to named-entity recognition problems, while 069 is due to
the lack of knowledge of “can travel freely within” being equivalent to
“have the right to live in”. The other 5 errors are all due to failure
of handling interation between quantifiers.


Former-commit-id: 218b420a14904b29916a665eb99d89e4b948d201