Implement all generalized quantifiers as Selections
This is just done for comparison. Compared to last commit when we
implement all generalized quantifiers as Relations, the performance on
the first section of FraCaS does not change. The system now gives
correct answer "unknown" for #030 and #032, as the Selection
implementation is less prone to false positives. However, the handling
of quantifier interaction is still flawed, as the system incorrectly
returns "unkown" for #031, #046, and #048. The new errors are on the data
points #044 and #076, due to failure of recoginizing downward entailment
on predicate argument of the quantifier.


Former-commit-id: aeeb8909e65e82d2a32c4986d7794726ce185a76